Select default

+++ msa/file_manager.py
@@ -26,6 +26,9 @@ def add_setting(alias, file_path):
 
 
 def activate_setting(file):
+    if not str(file):
+        return
+
     src_path = config.msa_path + file
     dst_path = config.m2_path + 'settings.xml'
 
@@ -35,6 +38,8 @@ def activate_setting(file):
 
 def deactivate_setting(file):
     src_path = config.m2_path + 'settings.xml'
+    if not os.path.exists(src_path):
+        return
 
     log.debug('Deactivate {}'.format(file))
     os.remove(src_path)
